Initial file upload
These are the files I've been working with before creating this repo.

The controller/main are self-explanatory to anyone who is familiar with JavaFx.

The sample.fxml is the main fxml template. Horiz1, horiz2, and wordbox are templates that are used by the add contextpack, wordpack, wordform etc. buttons. They are essential to the functionality of the project.